Added onBackpressed to ManageVehiclesActivity which forwards call on to ManageVehiclesFragment to clear the search text on Back key press before doing the default behavior (leaving the Activity in this case). Added a null argument check to the datasource for deleting Vehicle. Noted a possibly automatically added change to misc.xml, which seems to be a gradle build dependency.  it changed JDK from 1.8 back to 1.7 for some reason.

+++ app/src/main/java/com/airanza/vehicleservicehistory/VehicleServiceHistoryDataSource.java
@@ -151,6 +151,10 @@ public class VehicleServiceHistoryDataSource {
     }
 
     public void deleteVehicle(Vehicle vehicle) {
+        if(vehicle == null) {
+            return;
+        }
+
         long id = vehicle.getID();
         int i = database.delete(VehicleServiceHistoryDBHelper.VEHICLE_TABLE_NAME, VehicleServiceHistoryDBHelper.COLUMN_ID + " = " + id, null);
         if (i != 1) {
